Follow Jourdan's feedback updated
done : Is it possible to place the left and right videos centered in the white
space? not touching the edge of the window?

done : Can you make all of the body text the font Quicksand? Like the font on
the top banner.

done : In the parent testimonial section, can you lower the testimonial box so
it's not overlapping the video as much? Can those corners be rounded?
Thanks!

done : updated friend's image.